code: изменен тип данных с int на double

+++ main.cpp
@@ -4,9 +4,9 @@ using namespace std;
 
 int main(){
     cout << "Enter A and B: ";
-    int a, b;
+    double a, b;
     cin >> a >> b;
-    int max = a;
+    double max = a;
     if (b > a) max = b;
     cout << "a + b = "<< a + b << '\n'
     << "a - b = "<< a - b << '\n'
